+++ src/giskard_hub/data/chat_test_case.py
@@ -1,43 +1,43 @@
 from __future__ import annotations
 
 from dataclasses import dataclass, field
-from typing import Any, Dict, List, Optional
+from typing import Any, Dict, List, Optional, Union
 
-from ._base import BaseData
+from giskard_hub.data.check import CheckConfig, TestCaseCheckConfig
+
 from ._entity import Entity
 from .chat import ChatMessage, ChatMessageWithMetadata
 
 
-def _format_checks_to_cli(checks: List[TestCaseCheckConfig]) -> List[CheckConfig]:
+def _format_checks_to_cli(
+    checks: List[Union[TestCaseCheckConfig, Dict[str, Any]]],
+) -> List[CheckConfig]:
+    if not checks:
+        return []
+
+    checks = [check if isinstance(check, dict) else check.to_dict() for check in checks]
+
     return [
-        {
-            "identifier": check["identifier"],
-            "enabled": check["enabled"],
-            **(
-                {"params": params}
-                if check.get("assertions")
-                and (
-                    params := {
-                        k: v for k, v in check["assertions"][0].items() if k != "type"
-                    }
-                )
-                else {}
-            ),
-        }
+        CheckConfig.from_dict(
+            {
+                "identifier": check["identifier"],
+                "enabled": check["enabled"],
+                **(
+                    {"params": params}
+                    if check.get("assertions")
+                    and (
+                        params := {
+                            k: v
+                            for k, v in check["assertions"][0].items()
+                            if k != "type"
+                        }
+                    )
+                    else {}
+                ),
+            }
+        )
         for check in checks
     ]
-
-
-@dataclass
-class CheckConfig(BaseData):
-    identifier: str
-    params: Optional[dict[str, Any]] = None
-
-
-@dataclass
-class TestCaseCheckConfig(BaseData):
-    identifier: str
-    assertions: List[dict[str, Any]]
 
 
 @dataclass
