+++ src/giskard_hub/resources/_utils.py
@@ -1,4 +1,4 @@
-from typing import Any, List, Optional
+from typing import Any, Dict, List, Optional, Union
 
 from ..data._base import NOT_GIVEN, filter_not_given, maybe_to_dict
 from ..data.chat import ChatMessage, ChatMessageWithMetadata
@@ -30,15 +30,25 @@ def prepare_chat_test_case_data(
     )
 
 
-def _format_checks_to_backend(checks: list[CheckConfig]) -> list[TestCaseCheckConfig]:
+def _format_checks_to_backend(
+    checks: List[Union[CheckConfig, Dict[str, Any]]],
+) -> List[TestCaseCheckConfig]:
+    if not checks:
+        return []
+
+    checks = [check if isinstance(check, dict) else check.to_dict() for check in checks]
+
     return [
-        {
-            **check,
-            **(
-                {"assertions": [{"type": check["identifier"], **check["params"]}]}
-                if check.get("params")
-                else {}
-            ),
-        }
+        TestCaseCheckConfig.from_dict(
+            {
+                "enabled": True,  # Default value for enabled
+                **check,
+                **(
+                    {"assertions": [{"type": check["identifier"], **check["params"]}]}
+                    if check.get("params")
+                    else {}
+                ),
+            }
+        )
         for check in checks
     ]
